# Mammoth_Labs/blockscan-api/internal/adapter/http/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Backend Template",
        "title": "blockscan-go API",
        "contact": {},
        "license": {
            "name": "Go Echo"
        },
        "version": "1.0"
    },
    "host": "tracking.gmmtchain.io",
    "basePath": "/api/v1",
    "paths": {
...
        "/contract/wallet/nft/{wallet_address}": {
            "get": {
                "description": "ContractGetWalletNFTsByCollection / 월렛이 보유한 NFT들을 컬렉션 별로 분류해서 반환하는 API / *공용",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contract"
                ],
                "summary": "Get wallet nfts by collection / *공용",
                "operationId": "ContractGetWalletNFTsByCollection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x api key",
                        "name": "X-Api-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "지갑주소",
                        "name": "wallet_address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "블록체인 체인넘버",
                        "name": "chain_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "가져올 아이템 수 최대 100",
                        "name": "take",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "건너뛸 아이템 수 (옵션) 기본 0",
                        "name": "skip",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/blockscan-go_internal_adapter_http_response.ApiResponseSwagger-array_blockscan-go_internal_core_domain_contract_GetContractWalletNFTsByCollectionData"
                        }
                    }
                }
            }
        },
        "/contract/{hash}/wallet/nft/{wallet_address}": {
            "get": {
                "description": "ContractGetCollectionNFTsForWallet / 해당 컨트랙트에 월렛이 보유한 NFT들을 반환하는 API / *공용",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contract"
                ],
                "summary": "Get collection nfts for wallet / *공용",
                "operationId": "ContractGetCollectionNFTsForWallet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x api key",
                        "name": "X-Api-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "컨트랙트 주소",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "지갑주소",
                        "name": "wallet_address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "블록체인 체인넘버",
                        "name": "chain_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "가져올 아이템 수 최대 100",
                        "name": "take",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "건너뛸 아이템 수 (옵션) 기본 0",
                        "name": "skip",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/blockscan-go_internal_adapter_http_response.ApiResponseSwagger-blockscan-go_internal_core_domain_contract_GetContractCollectionNFTsForWalletData"
                        }
                    }
                }
            }
        },
...
        "/nft/detail/{hash}/{token_id}": {
            "get": {
                "description": "NftGetDetail / NFT 상세정보를 반환하는 API / *공용",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Nft"
                ],
                "summary": "Get Nft Detail / *공용",
                "operationId": "NftGetDetail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x api key",
                        "name": "X-Api-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "컨트랙트 주소",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "토큰 아이디",
                        "name": "token_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "블록체인 체인넘버",
                        "name": "chain_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/blockscan-go_internal_adapter_http_response.ApiResponseSwagger-blockscan-go_internal_core_domain_nft_GetNFtDetailData"
                        }
                    }
                }
            }
        },
        "/nft/detail/{hash}/{token_id}/{wallet_address}": {
            "get": {
                "description": "NftGetDetailOfWallet / 지갑이 보유한 NFT 상세정보를 반환하는 API / *공용",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Nft"
                ],
                "summary": "Get Nft Detail of Wallet / *공용",
                "operationId": "NftGetDetailOfWallet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x api key",
                        "name": "X-Api-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "컨트랙트 주소",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "토큰 아이디",
                        "name": "token_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "지갑주소",
                        "name": "wallet_address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "블록체인 체인넘버",
                        "name": "chain_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/blockscan-go_internal_adapter_http_response.ApiResponseSwagger-blockscan-go_internal_core_domain_nft_GetNFtDetailData"
                        }
                    }
                }
            }
        },
        "/nft/{wallet_address}": {
            "get": {
                "description": "NftGet / 월렛이 보유한 NFT들을 반환하는 API / *공용",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Nft"
                ],
                "summary": "Get Nfts by wallet / *공용",
                "operationId": "NftGet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x api key",
                        "name": "X-Api-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "지갑주소",
                        "name": "wallet_address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "블록체인 체인넘버",
                        "name": "chain_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "가져올 아이템 수 최대 100",
                        "name": "take",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "건너뛸 아이템 수 (옵션) 기본 0",
                        "name": "skip",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/blockscan-go_internal_adapter_http_response.ApiResponseSwagger-array_blockscan-go_internal_core_domain_nft_GetNftData"
                        }
                    }
                }
            }
        },
        "/transaction": {
            "post": {
                "description": "GetTransaction / 해당 지갑에서 코인/토큰 전송내역 가져오는 API / *공용",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "get coin, token transactions / *공용",
                "operationId": "GetTransaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x api key",
                        "name": "X-Api-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/blockscan-go_internal_core_domain_transaction.GetTransactionInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/blockscan-go_internal_adapter_http_response.ApiResponseSwagger-array_blockscan-go_internal_core_domain_transaction_GetTransactionData"
                        }
                    }
                }
            }
        },
        "/transaction/recent/{wallet_address}": {
            "get": {
                "description": "GetAllTransaction / 해당 지갑에서 코인/토큰 최근 전송내역 가져오는 API / *공용",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "get coin,token recent transactions / *공용",
                "operationId": "GetAllTransaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x api key",
                        "name": "X-Api-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "지갑주소",
                        "name": "wallet_address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "블록체인 넘버",
                        "name": "chain_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "가져올 아이템 수 최대 100",
                        "name": "take",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "건너뛸 아이템 수 (옵션) 기본 0",
                        "name": "skip",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/blockscan-go_internal_adapter_http_response.ApiResponseSwagger-array_blockscan-go_internal_core_domain_transaction_GetTransactionData"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "blockscan-go_internal_adapter_http_response.ApiResponseSwagger-array_blockscan-go_internal_core_domain_contract_GetContractWalletNFTsByCollectionData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/blockscan-go_internal_core_domain_contract.GetContractWalletNFTsByCollectionData"
                    }
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "$ref": "#/definitions/blockscan-go_internal_core_common_utils.ErrorType"
                }
            }
        },
...
        "blockscan-go_internal_adapter_http_response.ApiResponseSwagger-array_blockscan-go_internal_core_domain_nft_GetNftData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/blockscan-go_internal_core_domain_nft.GetNftData"
                    }
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "$ref": "#/definitions/blockscan-go_internal_core_common_utils.ErrorType"
                }
            }
        },
        "blockscan-go_internal_adapter_http_response.ApiResponseSwagger-array_blockscan-go_internal_core_domain_transaction_GetTransactionData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/blockscan-go_internal_core_domain_transaction.GetTransactionData"
                    }
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "$ref": "#/definitions/blockscan-go_internal_core_common_utils.ErrorType"
                }
            }
        },
...
        "blockscan-go_internal_adapter_http_response.ApiResponseSwagger-blockscan-go_internal_core_domain_contract_GetContractCollectionNFTsForWalletData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/blockscan-go_internal_core_domain_contract.GetContractCollectionNFTsForWalletData"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "$ref": "#/definitions/blockscan-go_internal_core_common_utils.ErrorType"
                }
            }
        },
...
        "blockscan-go_internal_adapter_http_response.ApiResponseSwagger-blockscan-go_internal_core_domain_nft_GetNFtDetailData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/blockscan-go_internal_core_domain_nft.GetNFtDetailData"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "$ref": "#/definitions/blockscan-go_internal_core_common_utils.ErrorType"
                }
            }
        },
...
        "blockscan-go_internal_core_common_utils.ErrorType": {
            "type": "string",
            "enum": [
                "No data found",
                "Syntax",
                "Duplicate",
                "Invalid sql statement",
                "Bad connection",
                "Invalid character",
                "Foreign key",
                "Check violation",
                "Success",
                "Fail",
                "Not found",
                "Invalid input",
                "Database",
                "Internal"
            ],
            "x-enum-varnames": [
                "NoDataFound",
                "SyntaxError",
                "DuplicateError",
                "InvalidSQLStatementError",
                "BadConnectionError",
                "JsonParseError",
                "ForeignKeyError",
                "CheckViolationError",
                "Success",
                "Fail",
                "NotFound",
                "InvalidInput",
                "DatabaseError",
                "InternalError"
            ]
        },
...
        "blockscan-go_internal_core_domain_contract.GetContractCollectionNFTsForWalletData": {
            "type": "object",
            "properties": {
                "aws_banner_image": {
                    "type": "string"
                },
                "aws_logo_image": {
                    "type": "string"
                },
                "chain_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "erc_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/blockscan-go_internal_core_domain_contract.ErcList"
                    }
                },
                "erc_type": {
                    "type": "string"
                },
                "hash": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                },
                "wallet_address": {
                    "type": "string"
                },
                "wallet_nickname": {
                    "type": "string"
                },
                "wallet_profile": {
                    "type": "string"
                }
            }
        },
        "blockscan-go_internal_core_domain_contract.GetContractWalletNFTsByCollectionData": {
            "type": "object",
            "properties": {
                "aws_banner_image": {
                    "type": "string"
                },
                "aws_logo_image": {
                    "type": "string"
                },
                "chain_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "erc_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/blockscan-go_internal_core_domain_contract.ErcList"
                    }
                },
                "erc_type": {
                    "type": "string"
                },
                "hash": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                },
                "tokens_amount": {
                    "type": "string"
                },
                "unique_tokens_count": {
                    "type": "string"
                },
                "wallet_address": {
                    "type": "string"
                },
                "wallet_nickname": {
                    "type": "string"
                },
                "wallet_profile": {
                    "type": "string"
                }
            }
        },
...
        "blockscan-go_internal_core_domain_nft.GetNFtDetailData": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "attributes_array": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/blockscan-go_internal_core_domain_nft.AttributesPercent"
                    }
                },
                "aws_banner_image": {
                    "type": "string"
                },
                "aws_logo_image": {
                    "type": "string"
                },
                "chain_id": {
                    "type": "integer"
                },
                "contract_description": {
                    "type": "string"
                },
                "contract_hash": {
                    "type": "string"
                },
                "contract_id": {
                    "type": "integer"
                },
                "contract_name": {
                    "type": "string"
                },
                "contract_symbol": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "nft_description": {
                    "type": "string"
                },
                "nft_name": {
                    "type": "string"
                },
                "token_id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "blockscan-go_internal_core_domain_nft.GetNftData": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "attributes_array": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/blockscan-go_internal_core_domain_nft.Attributes"
                    }
                },
                "aws_banner_image": {
                    "type": "string"
                },
                "aws_logo_image": {
                    "type": "string"
                },
                "chain_id": {
                    "type": "integer"
                },
                "contract_description": {
                    "type": "string"
                },
                "contract_hash": {
                    "type": "string"
                },
                "contract_id": {
                    "type": "integer"
                },
                "contract_name": {
                    "type": "string"
                },
                "contract_symbol": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "nft_description": {
                    "type": "string"
                },
                "nft_name": {
                    "type": "string"
                },
                "token_id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "blockscan-go_internal_core_domain_transaction.GetTransactionInput": {
            "type": "object",
            "required": [
                "chain_id",
                "wallet_address"
            ],
            "properties": {
                "chain_id": {
                    "type": "integer"
                },
                "chain_type": {
                    "type": "string",
                    "enum": [
                        "coin",
                        "token"
                    ]
                },
                "contract_address": {
                    "type": "string"
                },
                "order_by": {
                    "type": "string",
                    "enum": [
                        "asc",
                        "desc"
                    ]
                },
                "skip": {
                    "type": "integer"
                },
                "take": {
                    "type": "integer",
                    "minimum": 0
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "send",
                        "received"
                    ]
                },
                "wallet_address": {
                    "type": "string"
                }
            }
        }
    }
}